# swagger.json
{
    "swagger": "2.0",
    "info": {
        "version": "1.0.0",
        "title": "Progetto Gruppo T31 - IS2022",
        "description": "API del progetto del Gruppo T31 svolto durente il corso di Ingegneria del software - Anno Accademico 2022/23",
        "contact": {
            "name": "Luca Dematté - Referente gruppo T31",
            "email": "luca.dematte-2@studenti.unitn.it"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        }
    },
    "host": "gruppo-t31-api.onrender.com",
    "basePath": "/",
...
    "schemes": ["http", "https"],
    "paths": {
...
        "/ricorrenza/byID": {
            "get": {
                "tags": ["ricorrenza"],
                "summary": "Ottenere una ricorrenza tramite il suo ID",
                "description": "",
                "operationId": "getRicorrenzaConID",
                "consumes": [],
                "produces": ["application/json"],
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "description": "ID della ricorrenza da restituire",
                        "type": "string",
                        "required": true
                    },
                    {
                        "name": "x-access-token",
                        "in": "header",
                        "description": "Token di accesso dell'utente",
                        "type": "string",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": { "$ref": "#/definitions/Ricorrenza" }
                    },
                    "401": {
                        "description": "Unauthorized - Per svolgere l'operazione è necessario il login",
                        "schema": { "$ref": "#/definitions/ErroreAPI" }
                    },
                    "403": {
                        "description": "Forbidden - Il livello di privilegio non è sufficiente",
                        "schema": { "$ref": "#/definitions/ErroreAPI" }
                    },
                    "404": {
                        "description": "Not found - Una ricorrenza con l'ID fornito non esiste",
                        "schema": { "$ref": "#/definitions/ErroreAPI" }
                    },
                    "500": {
                        "description": "Internal server error - Errore imprevisto durante l'operazione",
                        "schema": { "$ref": "#/definitions/ErroreAPI" }
                    }
                }
            }
        },
        "/ricorrenza/byPeriodo": {
            "get": {
                "tags": ["ricorrenza"],
                "summary": "Specificando una data di inizio e di fine, viene restituita la lista di ricorrenze comprese nel periodo specificato",
                "description": "",
                "operationId": "getRicorrenzePerPeriodo",
                "consumes": [],
                "produces": ["application/json"],
                "parameters": [
                    {
                        "name": "inizio",
                        "in": "query",
                        "description": "Data di inizio del periodo",
                        "type": "string",
                        "format": "date-time",
                        "required": true
                    },
                    {
                        "name": "fine",
                        "in": "query",
                        "description": "Data di fine del periodo",
                        "type": "string",
                        "format": "date-time",
                        "required": true
                    },
                    {
                        "name": "x-access-token",
                        "in": "header",
                        "description": "Token di accesso dell'utente",
                        "type": "string",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": { "$ref": "#/definitions/Ricorrenza" }
                        }
                    },
                    "400": {
                        "description": "Bad request - I parametri richiesti non sono specificati o non sono coerenti",
                        "schema": { "$ref": "#/definitions/ErroreAPI" }
                    },
                    "401": {
                        "description": "Unauthorized - Per svolgere l'operazione è necessario il login",
                        "schema": { "$ref": "#/definitions/ErroreAPI" }
                    },
                    "403": {
                        "description": "Forbidden - Il livello di privilegio non è sufficiente",
                        "schema": { "$ref": "#/definitions/ErroreAPI" }
                    },
                    "500": {
                        "description": "Internal server error - Errore imprevisto durante l'operazione",
                        "schema": { "$ref": "#/definitions/ErroreAPI" }
                    }
                }
            }
        },
        "/ricorrenza/{id}": {
            "patch": {
                "tags": ["ricorrenza"],
                "summary": "Modifica di una ricorrenza",
                "description": "",
                "operationId": "modificaRicorrenza",
                "consumes": ["multipart/form-data"],
                "produces": ["application/json"],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID della ricorrenza da modificare",
                        "type": "string",
                        "required": true
                    },
                    {
                        "name": "x-access-token",
                        "in": "header",
                        "description": "Token di accesso dell'utente",
                        "type": "string",
                        "required": true
                    },
                    {
                        "name": "inizio",
                        "in": "formData",
                        "description": "Data e ora di inizio della ricorrenza",
                        "type": "string",
                        "format": "date-time",
                        "required": true
                    },
                    {
                        "name": "fine",
                        "in": "formData",
                        "description": "Data e ora di fine della ricorrenza",
                        "type": "string",
                        "format": "date-time",
                        "required": true
                    },
                    {
                        "name": "spaziPrenotati",
                        "in": "formData",
                        "description": "Lista di ID degli spazi utilizzati in questa ricorrenza",
                        "type": "array",
                        "items": { "type": "string" },
                        "required": true
                    },
                    {
                        "name": "serviziPrenotati",
                        "in": "formData",
                        "description": "Lista di ID dei servizi utilizzati in questa ricorrenza",
                        "type": "array",
                        "items": { "type": "string" },
                        "required": false
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK - Ricorrenza modificata correttamente",
                        "schema": { "$ref": "#/definitions/Ricorrenza" }
                    },
                    "400": {
                        "description": "Bad request - I parametri richiesti non sono specificati o non sono coerenti",
                        "schema": { "$ref": "#/definitions/ErroreAPI" }
                    },
                    "401": {
                        "description": "Unauthorized - Per svolgere l'operazione è necessario il login",
                        "schema": { "$ref": "#/definitions/ErroreAPI" }
                    },
                    "403": {
                        "description": "Forbidden - Il livello di privilegio non è sufficiente",
                        "schema": { "$ref": "#/definitions/ErroreAPI" }
                    },
                    "404": {
                        "description": "Not found - Una ricorrenza con l'ID fornito non esiste",
                        "schema": { "$ref": "#/definitions/ErroreAPI" }
                    },
                    "500": {
                        "description": "Internal server error - Errore imprevisto durante l'operazione",
                        "schema": { "$ref": "#/definitions/ErroreAPI" }
                    }
                }
            },
            "delete": {
                "tags": ["ricorrenza"],
                "summary": "Eliminazione di una ricorrenza",
                "description": "",
                "operationId": "eliminaRicorrenza",
                "consumes": [],
                "produces": ["application/json"],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID della ricorrenza da eliminare",
                        "type": "string",
                        "required": true
                    },
                    {
                        "name": "x-access-token",
                        "in": "header",
                        "description": "Token di accesso dell'utente",
                        "type": "string",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK - Eliminazione della ricorrenza effettuata correttamente",
                        "schema": { "$ref": "#/definitions/Ricorrenza" }
                    },
                    "401": {
                        "description": "Unauthorized - Per svolgere l'operazione è necessario il login",
                        "schema": { "$ref": "#/definitions/ErroreAPI" }
                    },
                    "403": {
                        "description": "Forbidden - Il livello di privilegio non è sufficiente",
                        "schema": { "$ref": "#/definitions/ErroreAPI" }
                    },
                    "404": {
                        "description": "Not found - Una ricorrenza con l'ID fornito non esiste",
                        "schema": { "$ref": "#/definitions/ErroreAPI" }
                    },
                    "500": {
                        "description": "Internal server error - Errore imprevisto durante l'operazione",
                        "schema": { "$ref": "#/definitions/ErroreAPI" }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "Spazio": {
            "type": "object",
            "required": ["_id", "nome", "tipologia", "prezzoIniziale", "prezzoOra"],
            "properties": {
                "_id": {
                    "type": "string",
                    "description": "Identificativo univoco dello spazio"
                },
                "nome": {
                    "type": "string",
                    "description": "Nome dello spazio",
                    "example": "Sala polifunzionale"
                },
                "descrizione": {
                    "type": "string",
                    "description": "Breve descrizione dello spazio",
                    "example": "Aula molto grande (100 posti) per spettacoli, conferenze, ..."
                },
                "tipologia": {
                    "type": "string",
                    "description": "Tipo di spazio",
                    "enum": ["Locale interno", "Locale attrezzato", "Spazio esterno"],
                    "example": "Locale interno"
                },
                "prezzoIniziale": {
                    "type": "number",
                    "format": "float",
                    "description": "Prezzo dello spazio quando viene utilizzato",
                    "example": 10.5
                },
                "prezzoOra": {
                    "type": "number",
                    "format": "float",
                    "description": "Prezzo dello spazio per ogni ora di utilizzo",
                    "example": 4.25
                },
                "URLfoto": {
                    "type": "string",
                    "description": "Percorso della foto dello spazio",
                    "example": "/images/spazi/default.png"
                },
                "servizi": {
                    "type": "array",
                    "description": "Lista dei servizi che possono essere utilizzati in questo spazio",
                    "items": { "$ref": "#/definitions/Servizio" },
                    "required": false
                }
            }
        },
        "Servizio": {
            "type": "object",
            "required": ["_id", "nome", "tipologia", "prezzoIniziale", "prezzoOra"],
            "properties": {
                "_id": {
                    "type": "string",
                    "description": "Identificativo univoco del servizio"
                },
                "nome": {
                    "type": "string",
                    "description": "Nome del servizio",
                    "example": "Proiettore"
                },
                "descrizione": {
                    "type": "string",
                    "description": "Breve descrizione del servizio",
                    "example": "Proiettore 720p con interfaccia HDMI e VGA"
                },
                "tipologia": {
                    "type": "string",
                    "description": "Tipo di servizio",
                    "enum": ["Arredamento", "Attrezzatura specifica", "Altro"],
                    "example": "Attrezzatura specifica"
                },
                "prezzoIniziale": {
                    "type": "number",
                    "format": "float",
                    "description": "Prezzo del servizio quando viene utilizzato",
                    "example": 2.5
                },
                "prezzoOra": {
                    "type": "number",
                    "format": "float",
                    "description": "Prezzo del servizio per ogni ora di utilizzo",
                    "example": 1.0
                },
                "URLfoto": {
                    "type": "string",
                    "description": "Percorso della foto del servizio",
                    "example": "/images/servizi/default.png"
                }
            }
        },
...
        "Ricorrenza": {
            "type": "object",
            "required": ["_id", "inizio", "fine", "spaziPrenotati"],
            "properties": {
                "_id": {
                    "type": "string",
                    "description": "Identificativo univoco della ricorrenza"
                },
                "inizio": {
                    "type": "string",
                    "format": "date-time",
                    "description": "Data e ora di inizio della ricorrenza",
                    "example": "2023-01-10T16:30:00.000Z"
                },
                "fine": {
                    "type": "string",
                    "format": "date-time",
                    "description": "Data e ora di fine della ricorrenza",
                    "example": "2023-01-10T18:30:00.000Z"
                },
                "spaziPrenotati": {
                    "type": "array",
                    "items": { "$ref": "#/definitions/Spazio" },
                    "description": "Lista degli spazi utilizzati in questa ricorrenza"
                },
                "serviziPrenotati": {
                    "type": "array",
                    "items": { "$ref": "#/definitions/Servizio" },
                    "description": "Lista dei servizi utilizzati in questa ricorrenza"
                }
            }
        },
        "ErroreAPI": {
            "type": "object",
            "required": ["code", "message"],
            "properties": {
                "code": {
                    "type": "integer",
                    "description": "Codice di risposta HTTP"
                },
                "message": {
                    "type": "string",
                    "description": "Messaggio di testo sull'errore"
                }
            }
        }
    }
}